# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/grade-levels/{gradeLevelId}/students/{studentId}": {
      "delete": {
        "operationId": "GradeLevelsController_unassignStudent",
        "parameters": [
          {
            "name": "gradeLevelId",
            "required": true,
            "in": "path",
            "description": "Grade level ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "responses": {
          "200": {
            "description": "Student unassigned"
          }
        },
        "summary": "Unassign a student from a grade level",
        "tags": [
          "Grade Levels"
        ]
      }
    },
...
    "/groups/{groupId}/students/{studentId}": {
      "delete": {
        "operationId": "GroupsController_unassignStudent",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "description": "Group ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "responses": {
          "200": {
            "description": "Student unassigned"
          }
        },
        "summary": "Unassign a student from a group",
        "tags": [
          "Groups"
        ]
      }
    },
...
    "/students": {
      "post": {
        "operationId": "StudentsController_create",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new student",
        "tags": [
          "students"
        ]
      },
      "get": {
        "operationId": "StudentsController_list",
        "parameters": [
          {
            "name": "centerId",
            "required": false,
            "in": "query",
            "description": "Filter by center ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "teacherId",
            "required": false,
            "in": "query",
            "description": "Filter by teacher ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "grade",
            "required": false,
            "in": "query",
            "description": "Filter by grade",
            "schema": {
              "type": "string",
              "enum": [
                "PRIMARY_1",
                "PRIMARY_2",
                "PRIMARY_3",
                "PRIMARY_4",
                "PRIMARY_5",
                "PRIMARY_6",
                "SECONDARY_1",
                "SECONDARY_2",
                "SECONDARY_3",
                "SECONDARY_4",
                "SECONDARY_5",
                "OTHER"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort by field",
            "schema": {
              "example": "name",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "example": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Filter by name (partial match)",
            "schema": {
              "example": "John",
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "students": [],
                    "total": 0,
                    "page": 1,
                    "limit": 10
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List students (by center, teacher, or all)",
        "tags": [
          "students"
        ]
      }
    },
    "/students/{id}": {
      "get": {
        "operationId": "StudentsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get student by ID",
        "tags": [
          "students"
        ]
      },
      "put": {
        "operationId": "StudentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update student by ID",
        "tags": [
          "students"
        ]
      },
      "delete": {
        "operationId": "StudentsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "responses": {
          "204": {
            "description": "Student deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete student by ID",
        "tags": [
          "students"
        ]
      }
    },
...
  },
  "info": {
    "title": "LMS SaaS API",
    "description": "API documentation for the LMS SaaS platform.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateStudentDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "User ID for the student"
          },
          "teacherId": {
            "type": "string",
            "format": "uuid",
            "description": "Teacher ID if student belongs to a freelance teacher"
          },
          "centerId": {
            "type": "string",
            "format": "uuid",
            "description": "Center ID if student belongs to a center"
          },
          "grade": {
            "type": "string",
            "enum": [
              "PRIMARY_1",
              "PRIMARY_2",
              "PRIMARY_3",
              "PRIMARY_4",
              "PRIMARY_5",
              "PRIMARY_6",
              "SECONDARY_1",
              "SECONDARY_2",
              "SECONDARY_3",
              "SECONDARY_4",
              "SECONDARY_5",
              "OTHER"
            ],
            "description": "Student grade (enum)"
          },
          "level": {
            "type": "string",
            "description": "Academic level or curriculum (e.g. IGCSE, American, National)"
          },
          "guardianId": {
            "type": "string",
            "format": "uuid",
            "description": "Guardian ID if linked"
          },
          "performanceScore": {
            "type": "number",
            "description": "Performance score (float)"
          },
          "totalSessionsAttended": {
            "type": "number",
            "description": "Total sessions attended",
            "default": 0
          },
          "totalPayments": {
            "type": "number",
            "description": "Total payments (float)",
            "default": 0
          },
          "notes": {
            "type": "string",
            "description": "Notes about the student"
          }
        },
        "required": [
          "userId",
          "grade"
        ]
      },
      "StudentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "teacherId": {
            "type": "string"
          },
          "centerId": {
            "type": "string"
          },
          "grade": {
            "type": "string",
            "enum": [
              "PRIMARY_1",
              "PRIMARY_2",
              "PRIMARY_3",
              "PRIMARY_4",
              "PRIMARY_5",
              "PRIMARY_6",
              "SECONDARY_1",
              "SECONDARY_2",
              "SECONDARY_3",
              "SECONDARY_4",
              "SECONDARY_5",
              "OTHER"
            ]
          },
          "level": {
            "type": "string"
          },
          "guardianId": {
            "type": "string"
          },
          "performanceScore": {
            "type": "number"
          },
          "totalSessionsAttended": {
            "type": "number"
          },
          "totalPayments": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "type": "object"
          },
          "teacher": {
            "type": "object"
          },
          "center": {
            "type": "object"
          },
          "guardian": {
            "type": "object"
          },
          "groups": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "id",
          "userId",
          "grade",
          "totalSessionsAttended",
          "totalPayments",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateStudentDto": {
        "type": "object",
        "properties": {
          "teacherId": {
            "type": "string",
            "format": "uuid",
            "description": "Teacher ID if student belongs to a freelance teacher"
          },
          "centerId": {
            "type": "string",
            "format": "uuid",
            "description": "Center ID if student belongs to a center"
          },
          "grade": {
            "type": "string",
            "enum": [
              "PRIMARY_1",
              "PRIMARY_2",
              "PRIMARY_3",
              "PRIMARY_4",
              "PRIMARY_5",
              "PRIMARY_6",
              "SECONDARY_1",
              "SECONDARY_2",
              "SECONDARY_3",
              "SECONDARY_4",
              "SECONDARY_5",
              "OTHER"
            ],
            "description": "Student grade (enum)"
          },
          "level": {
            "type": "string",
            "description": "Academic level or curriculum (e.g. IGCSE, American, National)"
          },
          "guardianId": {
            "type": "string",
            "format": "uuid",
            "description": "Guardian ID if linked"
          },
          "performanceScore": {
            "type": "number",
            "description": "Performance score (float)"
          },
          "totalSessionsAttended": {
            "type": "number",
            "description": "Total sessions attended",
            "default": 0
          },
          "totalPayments": {
            "type": "number",
            "description": "Total payments (float)",
            "default": 0
          },
          "notes": {
            "type": "string",
            "description": "Notes about the student"
          }
        }
      },
...
    }
  }
}